Move router.ts to utils/ directory
